@@ -1,4 +1,4 @@
 source 'https://rubygems.org'
-ruby "2.0.0"
-gem 'rspec','2.14.1'
+ruby '2.2.0'
+gem 'rspec'
 gem 'pry'
