@@ -5,7 +5,7 @@ Dir.glob("lib/*.rb").each do |lib|
   euler = eval lib.sub("lib/","").sub(".rb","").capitalize
   describe euler do
     it 'return collect answer' do
-      euler.solve.should == answers[euler.to_s[-3,3].to_i-1]
+      expect(euler.solve).to eq(answers[euler.to_s[-3,3].to_i-1])
     end
   end
 end
